xserver xkeyboard-config mesa putty libX11 mkfontscale fontconfig EGL-Registry xorgproto git update Dec 30 2021
xserver          commit 38291fa86c00e1d3ee37612c9215b0c15da89051
xkeyboard-config commit 0554e6c220555d1ed94b5ae6253226516faae951
libX11           commit 4c96f3567a8d045ee57b886fddc9618b71282530
mkfontscale      commit 2ee95eca936971e841a135ecc4f4a25875232fa2
fontconfig       commit e474b380ffdb330eca7dd0f0fddacc7b19f3247d
mesa             commit 71e2df73d95c418ff21312b0ed7b8420eb927999
putty            commit f9d0557330f985ff23dbe5b8e014dc22cecbf13e
EGL-Registry     commit 7c093e9891ee8c49ed2567ffd38c83d148ecbddc
xorgproto        commit 914d8f5e0f469cd0416364dd008e9eea752bf703

+++ mesalib/src/freedreno/decode/rnnutil.c
@@ -201,7 +201,7 @@ rnn_regoff(struct rnn *rnn, uint32_t offset)
 }
 
 enum rnnttype
-rnn_decodelem(struct rnn *rnn, struct rnntypeinfo *info, uint32_t regval,
+rnn_decodelem(struct rnn *rnn, struct rnntypeinfo *info, uint64_t regval,
               union rnndecval *val)
 {
    val->u = regval;
